Novel_Solutions/requirements.txt
flask
python-dotenv
flask_sqlalchemy